+++ packages/grida-cmath/__tests__/cmath.ext.snap.spacing.test.ts
@@ -1,15 +1,15 @@
-import { snap } from "../_snap";
+import { cmath } from "..";
 
 const sortnum = (a: number, b: number) => a - b;
 
 describe("snap.spacing.projection", () => {
   it("should calculate projections for non-overlapping ranges", () => {
-    const ranges: snap.spacing.Range[] = [
+    const ranges: cmath.Range[] = [
       [0, 10],
       [20, 30],
     ];
 
-    const result = snap.spacing.repeatedpoints(ranges);
+    const result = cmath.ext.snap.spacing.repeatedpoints(ranges);
 
     expect(result.a.flat()).toEqual([40]); // Projection based on b2 + space
     expect(result.b.flat()).toEqual([-10]); // Projection based on a1 - space
@@ -17,12 +17,12 @@ describe("snap.spacing.projection", () => {
   });
 
   it("should return empty projections for overlapping ranges", () => {
-    const ranges: snap.spacing.Range[] = [
+    const ranges: cmath.Range[] = [
       [0, 10],
       [5, 15],
     ];
 
-    const result = snap.spacing.repeatedpoints(ranges);
+    const result = cmath.ext.snap.spacing.repeatedpoints(ranges);
 
     expect(result.a.flat()).toEqual([]); // No valid projections
     expect(result.b.flat()).toEqual([]); // No valid projections
@@ -30,13 +30,13 @@ describe("snap.spacing.projection", () => {
   });
 
   it("should handle multiple non-overlapping ranges", () => {
-    const ranges: snap.spacing.Range[] = [
+    const ranges: cmath.Range[] = [
       [0, 10], // 1
       [20, 30], // 2
       [40, 50], // 3
     ];
 
-    const result = snap.spacing.repeatedpoints(ranges);
+    const result = cmath.ext.snap.spacing.repeatedpoints(ranges);
 
     expect(result.a).toEqual([
       [40], // 1_2
@@ -48,9 +48,9 @@ describe("snap.spacing.projection", () => {
   });
 
   it("should handle a single range (no combinations possible)", () => {
-    const ranges: snap.spacing.Range[] = [[0, 10]];
+    const ranges: cmath.Range[] = [[0, 10]];
 
-    const result = snap.spacing.repeatedpoints(ranges);
+    const result = cmath.ext.snap.spacing.repeatedpoints(ranges);
 
     expect(result.a.flat()).toEqual([]); // No combinations possible
     expect(result.b.flat()).toEqual([]); // No combinations possible
@@ -58,9 +58,9 @@ describe("snap.spacing.projection", () => {
   });
 
   it("should handle empty input", () => {
-    const ranges: snap.spacing.Range[] = [];
+    const ranges: cmath.Range[] = [];
 
-    const result = snap.spacing.repeatedpoints(ranges);
+    const result = cmath.ext.snap.spacing.repeatedpoints(ranges);
 
     expect(result.a.flat()).toEqual([]);
     expect(result.b.flat()).toEqual([]);
@@ -68,13 +68,13 @@ describe("snap.spacing.projection", () => {
   });
 
   it("should ignore invalid combinations (negative space)", () => {
-    const ranges: snap.spacing.Range[] = [
+    const ranges: cmath.Range[] = [
       [0, 10],
       [5, 15], // Overlapping range
       [20, 30],
     ];
 
-    const result = snap.spacing.repeatedpoints(ranges);
+    const result = cmath.ext.snap.spacing.repeatedpoints(ranges);
 
     expect(result.a.flat().sort(sortnum)).toEqual([35, 40]); // Only valid projection
     expect(result.b.flat().sort(sortnum)).toEqual([-10, 0]); // Only valid projection
@@ -82,7 +82,7 @@ describe("snap.spacing.projection", () => {
   });
 
   it("should handle multiple ranges with 5 inputs", () => {
-    const ranges: snap.spacing.Range[] = [
+    const ranges: cmath.Range[] = [
       [0, 10], // First range
       [15, 25], // Second range
       [30, 40], // Third range
@@ -90,7 +90,7 @@ describe("snap.spacing.projection", () => {
       [70, 80], // Fifth range
     ];
 
-    const result = snap.spacing.repeatedpoints(ranges);
+    const result = cmath.ext.snap.spacing.repeatedpoints(ranges);
 
     const sortnum = (a: number, b: number) => a - b;
